updating folder tree structure.
