+++ Dockerfile
@@ -2,7 +2,7 @@ FROM  python:3.10-slim-buster
 WORKDIR /app
 COPY ./requirements.txt .
 RUN pip install --no-cache-dir --upgrade -r requirements.txt
-COPY . .
-EXPOSE 5000
+COPY ./weatherapp .
+EXPOSE 8080
 
-CMD ["gunicorn", "wsgi:app", "-b", "0.0.0.0:5000"]
+CMD ["gunicorn", "wsgi:app", "-b", "0.0.0.0:8080"]
