+++ requirements.txt
@@ -1,5 +1,6 @@
 gunicorn==23.0.0
 Flask==3.0.1
 requests==2.23.0
+pillow==11.0.0 
 pyowm==3.0.0
 pyzipcode==3.0.1
