+++ weatherapp/templates/500.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Weather App</title>
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css">
+  <style>
+    body {
+      padding: 20px;
+    }
+    .weather-card {
+      max-width: 400px;
+      margin: 0 auto;
+      background-color: #f7f7f7;
+      padding: 20px;
+      border-radius: 5px;
+      box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
+    }
+    .weather-icon {
+            max-width: 100px;
+            margin: 0 auto;
+            display: block;
+    }
+    </style>
+    </head>
+    <body>
+        <div class="container">
+            <div class="weather-card">
+                <img src="{{url_for('static', filename='500.jpg')}}" height=500px, width=500px />
+            </div>
+        </div>
+    </body>
+</html>
+
